drivers: sensor: ina23x: specify current LSB in uA and fix units
Specify the units of the current LSB in microamperes, so that we can
measure low maximum currents. Right now it was specified in
milliamperes, but ignored and always hardcoded to 1mA in the driver.
This makes the driver pretty much useless when the maximum current to be
measured is in a range of e.g. 20-50mA.

This patch also removes some unnecessary ifdeffery: since we write the
calibration register, we can always provice measurements with the right
units. It is also wrong to provide sensor readings that do not match
with the units specified by the channel. After this change voltage is
always reported in V, current in A and power in W.

Note that power measurement had the current LSB hardcoded in the
calculation (assuming 1mA/LSB), this has been fixed as well.

+++ drivers/sensor/ina23x/ina237.h
@@ -48,7 +48,7 @@ struct ina237_config {
 	struct i2c_dt_spec bus;
 	uint16_t config;
 	uint16_t adc_config;
-	uint16_t current_lsb;
+	uint32_t current_lsb;
 	uint16_t rshunt;
 	const struct gpio_dt_spec gpio_alert;
 	uint16_t alert_config;
